- Removed blast from Cliquely, now only compares fasta sequences to the local db.
+Improved CliquesGraph memory usage
- Removed Eukaryotas.

+++ CliquesGraphs/Graphs/CliqueSizeGraph.cs
@@ -27,7 +27,7 @@ namespace CliquesGraphs.Graphs
         }
 
 
-        private int[] CountCliquesSizes(List<int> cliquesSizes)
+        private int[] CountCliquesSizes(List<ushort> cliquesSizes)
         {
             var maxCliqueSize = cliquesSizes.Max();
             var cliquesSizeCounter = new int[maxCliqueSize];
